// src/modules/contact/type.ts
import React from "react";

export type InputType = {
  name: string;
  email: string;
  message: string;
};
export type EmailType = {
  name: string;
  email: string;
  message: string;
};

export type EmailTempType = {
  name: string;
  email: string;
  message: string;
};

export type ContactAddressType = {
  title: string;
  details: string;
  Icon: React.ElementType;
};